Adding a lot changes
Generally speaking those change are awesome, I can't really pinpoint which one
is better cause I am so excited by all of this. It is indeed an exciting moment
and that's the way we should do it.

If there is one thing I learn is that terse commit are boring, but there is
another I learnt and that's it!

Have a wonderful day!